Handle no audio
When a Beep or Alarm instances can't be created, replace them with a NoSound stub instances. It overrides all overridable methods in ClipWrapper with empty method bodies. This way, no null checks are needed. Or, null checks are moved into ClipWrapper's methods.

If both, Beep and Alarm fail, volume related components are disabled.